Uploaded original source file
Supports the creation of a CMD object that allows for commands to be written to the command line and output to be read from the command line. There is no support for issuing the 'ctrl-c' keystroke to the command line.  There is no support for controlling a sub-process in the command window, Ex. Running the command 'wmic' will place the command windows in an uncontrollable state. The process is awaiting user input to the wmic process; however, we can only write to the cmd process which is currently not accepting input.